Added map and button to viewController. Map not shows current location and centers on current location every time a new map tile is finished loading (not intended behavior). Also logging lots of text at this point to figure out when certain events fire and in what order.
Added a PadGestureRecognizer and a TapGestureRecognizer to I can tell when the user pans the map. The idea behind that is I will be able to detect when I should stop tracking the users current location and when a user adds a new pin to the map.

Playing around with the idea of a profile class/struct so I can manage the user's collected data and start moving stuff out of the viewController. Not sure where that is going yet.

Added an Array and NSData extension class, but turns out I probably don't need those (for now). Will see where it goes.

Info.plist was updated to allow background updates and use of the gps and location services. I intend to grab the user's current location within, say 1 mile, and start building a heat map of where the user goes. That work is not done in this commit. Just documenting intent.

Currently getting a warning about signing is about to expire, or something like that. Not sure where that is coming from or how to fix it.